add(assets/draftProps.tsx) add common actions
