+++ src/components/articles/assets/index.css
@@ -1,0 +1,47 @@
+.header-one {
+  font-size: 2em !important;
+  color: #555;
+  margin: 3em 0 0 0 !important;
+  padding: .3em .5em !important;
+  background-color: #fcc;
+}
+
+.header-two {
+  font-weight: 600 !important;
+  font-size: 1.4em !important;
+  margin-top: 0 !important;
+  padding-left: .5em !important;
+  border-left: solid 5px #f55;
+}
+
+.header-three > div > span {
+  position: relative;
+  padding: 0.2em 0.3em;
+  margin: .5em 0 .1em 0 !important;
+  background: #fee;
+  border-left: 3px solid #f55 ;
+  /* border-bottom: 2px solid #f55 ; */
+  color: #555;
+  font-weight: bold;
+  font-size: .9em !important;
+} 
+
+.header-three > div > span:before {
+  content: '>';
+  color: #f55;
+  margin-right: .3em;
+}
+.paragraph {
+  border-left: 1px solid #f99;
+  margin-bottom: 1em !important;
+  padding-left: .5em !important;
+}
+
+.header-six {
+  color: #f55;
+  font-size: 1em !important;
+}
+
+.public-DraftStyleDefault-ul {
+  border-left: solid 1px #99f; 
+}
