+++ src/plugins/index.ts
@@ -1,0 +1,4 @@
+import Image from './image'
+import Table from './table'
+
+export default [ Image, Table ]
